Add explicit type conversions

+++ src/libscotch/vgraph_separate_df.c
@@ -1,4 +1,4 @@
-/* Copyright 2007,2013,2018,2021,2023 IPB, Universite de Bordeaux, INRIA & CNRS
+/* Copyright 2007,2013,2018,2021,2023,2024 IPB, Universite de Bordeaux, INRIA & CNRS
 **
 ** This file is part of the Scotch software package for static mapping,
 ** graph partitioning and sparse matrix ordering.
@@ -53,7 +53,7 @@
 /**                # Version 6.1  : from : 27 nov 2021     **/
 /**                                 to   : 27 nov 2021     **/
 /**                # Version 7.0  : from : 16 jan 2023     **/
-/**                                 to   : 16 jan 2023     **/
+/**                                 to   : 09 aug 2024     **/
 /**                                                        **/
 /************************************************************/
 
@@ -174,8 +174,8 @@ const VgraphSeparateDfParam * const paraptr)      /*+ Method parameters +*/
     difotax[grafptr->s.vertnnd - 1] = vanctab[1] / edlstax[grafptr->s.vertnnd - 1];
 
     veloval = 1.0F;                               /* Assume no vertex loads */
-    cdifval = paraptr->cdifval;
-    cremval = paraptr->cremval;
+    cdifval = (float) paraptr->cdifval;
+    cremval = (float) paraptr->cremval;
     for (movenum = 0; movenum < paraptr->movenbr; movenum ++) { /* For all moves */
       Gnum                vertnum;
       Gnum                vertnnd;
